docs: Splitting out to modules for doc gen
