+++ packages/example/package/src/index.ts
@@ -1,3 +1,13 @@
+/**
+ * @module Example
+ * @prefered
+ *
+ * This comment!
+ *  */ /** and now?*/
+export * from "./create-user";
+export * from "./user";
+
+/** Hey there! */
 export function helloWorld(): string {
 	return `Hello World!`;
 }
@@ -5,28 +15,3 @@ export function helloWorld(): string {
 export function echo(stringVal: string): string {
 	return stringVal;
 }
-
-/**
- * @param user The user to be created in the db
- * @returns The created user with the id populated
- */
-export async function createUser(user: User): Promise<User> {
-	return new Promise((resolve) => {
-		user.id = `1`;
-		resolve(user);
-	});
-}
-
-/**
- * This user _supports_ [Markdown](https://marked.js.org/)
- * @example
- * ```
- * const bob = { id:'1', firstName:'bob' }
- * ```
- */
-export type User = {
-	/** The unique identifier of the user */
-	id?: string;
-	/** The first name of the user */
-	firstName: string;
-};
