+++ packages/example/package/src/user.ts
@@ -1,0 +1,14 @@
+/** @module Example */
+/**
+ * This user _supports_ [Markdown](https://marked.js.org/)
+ * @example
+ * ```
+ * const bob = { id:'1', firstName:'bob' }
+ * ```
+ */
+export type User = {
+	/** The unique identifier of the user */
+	id?: string;
+	/** The first name of the user */
+	firstName: string;
+};
